Use custom schema for predictor events
