+++ neuroscout/models/predictor.py
@@ -1,6 +1,4 @@
 from database import db
-import statistics
-
 
 class Predictor(db.Model):
     """ Instantiation of a predictor in a dataset.
@@ -54,23 +52,3 @@ class PredictorRun(db.Model):
     run_id = db.Column(db.Integer, db.ForeignKey('run.id'), primary_key=True)
     predictor_id = db.Column(db.Integer, db.ForeignKey('predictor.id'),
                              primary_key=True)
-
-    def stat_property(function):
-        @property
-        def wrapper(self):
-            val_query = PredictorEvent.query.filter_by(
-                    run_id=self.run_id,
-                    predictor_id=self.predictor_id).with_entities('value')
-            try:
-                return function(self, [float(a[0]) for a in val_query])
-            except ValueError:
-                return None
-        return wrapper
-
-    @stat_property
-    def mean(self, values):
-        return statistics.mean(values)
-
-    @stat_property
-    def stdev(self, values):
-        return statistics.stdev(values)
